+++ test/string.js
@@ -189,36 +189,38 @@ describe('String', function() {
 	});
 
 
-	describe('Loose Number String', function() {
+	describe('Value String', function() {
 
 		it('should return false if not passed a string', function() {
-			expect(valid.numberStringLoose(0)).to.be.false;
+			expect(valid.valueString(0)).to.be.false;
 		});
 
 		it('should return true if passed a valid number string', function() {
-			expect(valid.numberStringLoose('0')).to.be.true;
-			expect(valid.numberStringLoose('-0')).to.be.true;
-			expect(valid.numberStringLoose('+0')).to.be.true;
-			expect(valid.numberStringLoose('123')).to.be.true;
-			expect(valid.numberStringLoose('-123')).to.be.true;
-			expect(valid.numberStringLoose('+123')).to.be.true;
-			expect(valid.numberStringLoose('123.123')).to.be.true;
-			expect(valid.numberStringLoose('123px')).to.be.true;
-			expect(valid.numberStringLoose('123%')).to.be.true;
-			expect(valid.numberStringLoose('$123')).to.be.true;
-			expect(valid.numberStringLoose('123-123')).to.be.true;
-			expect(valid.numberStringLoose('123,000')).to.be.true;
-			expect(valid.numberStringLoose('123,000.00')).to.be.true;
+			expect(valid.valueString('0')).to.be.true;
+			expect(valid.valueString('-0')).to.be.true;
+			expect(valid.valueString('+0')).to.be.true;
+			expect(valid.valueString('123')).to.be.true;
+			expect(valid.valueString('-123')).to.be.true;
+			expect(valid.valueString('+123')).to.be.true;
+			expect(valid.valueString('123.123')).to.be.true;
+			expect(valid.valueString('123px')).to.be.true;
+			expect(valid.valueString('123BTC')).to.be.true;
+			expect(valid.valueString('123 BTC')).to.be.true;
+			expect(valid.valueString('123%')).to.be.true;
+			expect(valid.valueString('$123')).to.be.true;
+			expect(valid.valueString('123-123')).to.be.true;
+			expect(valid.valueString('123,000')).to.be.true;
+			expect(valid.valueString('123,000.00')).to.be.true;
 		});
 
 		it('should return false if passed a bad string', function() {
-			expect(valid.numberStringLoose('')).to.be.false;
-			expect(valid.numberStringLoose('NaN')).to.be.false;
-			expect(valid.numberStringLoose('test')).to.be.false;
-			expect(valid.numberStringLoose('te123st')).to.be.false;
-			expect(valid.numberStringLoose('Infinity')).to.be.false;
-			expect(valid.numberStringLoose('-')).to.be.false;
-			expect(valid.numberStringLoose('+')).to.be.false;
+			expect(valid.valueString('')).to.be.false;
+			expect(valid.valueString('NaN')).to.be.false;
+			expect(valid.valueString('test')).to.be.false;
+			expect(valid.valueString('te123st')).to.be.false;
+			expect(valid.valueString('Infinity')).to.be.false;
+			expect(valid.valueString('-')).to.be.false;
+			expect(valid.valueString('+')).to.be.false;
 		});
 
 	});